Management of external remote kernels in ExternalIPython: * ExternalIPython can now use kernels started by an independent (3rd party) process such as jupyter notebook (I call these kernels "remote" even if they are run on the local machine).
Using a few hacks in the consoles.ExternalIPython, consoles.ExternalConsoleWindow
and workspacemodel.WorkspaceModel classes, the symbols present in a
remote kernel's namespace at the time when the External Console session
was ended are now visible again in Scipyen's Workspace table ("User variables")
when a new session with the kernel is started, provided the kernel was kept alive in the meanwhile.

The process takes into account any changes brought to the remote kernel
workspace by the 3rd party, independent, program that started the
remote kernel.

The list of symbols is saved in the "cached_sessions.json" file saved
in the Scipyen's user configuration directory (on Linux, this is
$HOME/.config/Scipyen)



Signed-off-by: Cezar M. Tigaret <cezar.tigaret@gmail.com>