Cursors have now a __call__ method as well; this generates a QPainterPath that is used in calculating the shape of the GraphicsObject, for detecting mouse hits, collisions etc.
This now, finally, allows cursors to be dragged from their main lines around
the scene (image) in the viewer.

Things seem to work 99.999% OK.

The bug related to the offset might be related to the ImageViewer displaying
a message when a new cursor is being added to the scene via the GUI (which
causes a shift of the scene on screen) - might need to revisit this.

Next steps: cursor behaviour with multi-frame backends, and synchronizing
several frontends for the same backend.



Signed-off-by: Cezar M. Tigaret <cezar@w009.psych.cf.ac.uk>