Continuing the overhaul of PlanarGraphics/GraphicsObjects framework: In ImageViewer/GraphicsImageViewerWidget move away from the (convoluted) management of graphics objects and planar graphics objects (cased on chainMap etc). This i sbased on the fact that the GraphicsImageViewerWidget scene (which is a GraphicsImageViewerScene) already stores the GraphicsObjects - there is not point in yet another collection of references to these objects. In turn, every GraphicsObject object stored a reference to its own backend (the PlanarGraphics). So the access to a PlanarGraphics object is made now by directly accessing the 'backend' attribute of a GraphicsObject object that exists in the scene.
This is done through the use of generators and iterators that also allow
selecting a PlanarGraphics object by the values of its attributes.

The generator and iterator themselves are factored out in core.prog.py
module.

TODO: clean up the code in imageviewer.py module, then get back to the
overhaul of the PlanarGraphics & GraphicsObject classes for non-cursor
types.

Signed-off-by: Cezar M. Tigaret <cezar.tigaret@gmail.com> 09/05/2021