Gradient editing actions (hover poiint move/add/remove in shade widgets) accomplished, for better consistency with gradient stops editing:
* moving a hover point no longer generates new gradient stops; instead the
 X coordinate  of the corresponding hove points in the non-interacted shade
widgets are updated to reflect the (changed) position of the gradient stop

* adding/removing a hover point in a shade widgets inserts hover points at
the same X coordinate in the other shade widgets, or, respectively, removes
the corresponding hover point in the other shade widgets.

This solves a design limitation in the Qt gradients example which was not
intended to be used as a real-life gradient editor.

Signed-off-by: Cezar M. Tigaret <cezar.tigaret@gmail.com>