Channel calibration data: "index" is not necessarily the running index of the channel. For instance, PrairieView sets it up to 1-based running index (int stored as str).
For this reason (and to faithfully represent the channel index set by
the acquisition software) channel index is allowed to take an arbitrary
int value.

As a consequence, the index parameters in
 AxisCalibrationData.getChannelCalibration(index, full)

can now be:
*  a str: either the symbol to which the channel calibration data
is mapped in the AxisCalibrartionData object or the value of the
'name' parameter in the ChannelCalibration objects
* an int (the value of the 'index' parameter of the ChannelCalibration
objects)
* None (in which case the function returns the first available
ChannelCalibration in the AxisCalibrationData object of a Channels axis.


Signed-off-by: Cezar M. Tigaret <cezar.tigaret@gmail.com>