fixed a bug due to comparing pandas.NA with a value, in core.multiframeindex
removes scipyen_desktop.sh; I now use the same script (scipyen) for
launching scipyen from the terminal, or via a desktop file (Linux)

envact: script which defined the 'scipyact' shell function which activates
the python virtual environment for running scipyen and is sourced by
the scipyen script. This ensures that scipyen can be called to launch Scipyen
without having to activate the environment specifically (on can stil do it
like this, but now one can type a single command)

Upon first installation, or changing the python virtual environment under
which Scipyen is supposed ot run, the script 'envact' must be edited
manually so that the correct environment is activated.

To activate the virtual environment WITHOUT launching Scipyen, the 'envact'
script file also needs to be sourced in (or its contents copied to) the
'.bashrc' file (or one of '.zshrc', '.cshrc', '.tcshrc' depending on what
shell you actually use - read their corresponding UNIX man pages for more details)






The script 'scipyen' needs to be made executable and a symbolic link to it created
in a directory included in the PATH environment variable (usually, $HOME/bin)
the preliminary invocation of the environment's activate script.

Signed-off-by: Cezar M. Tigaret <cezar.tigaret@gmail.com>