Finally storing configuration via confuse seems to work fine; some bugs from the traitutils and traitcontainers linked to managinf non-Qt confugrables cleared.
The framework still needs work/cleaning up/ removing reducnant code but
it basically seems to do what is proposes to do.

Back to mPSCAnalysis, finalize:
1) storing template waveform to file, reloading it if needed
1.1) test feasibility of using axon text files with signal data - useful
for other apps as well
2) implement the mPSC detection & analysis per sweep or per entire data

Then move on to LTP

Then go back and finalize ScanData - Analysis Units, cursors and regions,
and LSCaT

Write/improve documentation alongside to that.

Target: prepare to an alpha release.

Signed-off-by: ctigaret <cezar.tigaret@gmail.com>