More progress on plugins framework.
Apps can be launched as plugins and (hopefully) there are no clashes with already loaded modules
(in fact a plugin can be loaded manually and the plugin loading logic should
already detect that)

TODO: 22k23
When appropriate, make plugins advertise themselves as applications independently
of their provided menu.

Also, when appropriate, make plugins advertise themselves as viewers to be registered
with gui viewers in main window (this, alsoindependent of their provided
GUI menu).

Signed-off-by: ctigaret <cezar.tigaret@gmail.com>