OK: 1. local batch file input not blocking anymore. 2. workspace viewer updates more swiftly, and not UI blocking, when there are many variables monitored But the traitcontainer update still scales up with the number of objects (ideally, it should be constant)
FIXME:
When loading a large batch of files, each file load
triggers a workspace model update; this can still take longer
than comfortable, although the UI is not blocking.
The fact that the whole loop is monitored with a progress dialog
and can be cancelled is OK.

Possible workaround: an option where the batch file loading
is decoupled from the workspace viewer update (and make the latter update
in a loop inside a separate thread AFTER all file have been loaded).

Currently, the second stage blocks the UI -- FIXME

Signed-off-by: Cezar Tigaret <cezar.tigaret@gmail.com>