Okay: using a startup splash is less trivial than I though:
1) on its own, it does not populate itself until after the main window
 is shown

2) passed to the main window constructor and updating it from within mainwindow.__init__
works well (the splash is populated and new messages appear as they are
being set), BUT this messes up the mainwindow's menubar when it is set
as NativeMenuBar

I.e., the mainwindow's menu bar is not being shown as a global menu.

Manually disabling the UseNativeMenuBar property shows the menu bar in its
window.

So the problem is:

how can I show a dynamic splash with information updates as the mainwindow
initializes itself, AND enable the native menu bar in the said main window?

Until I find a solution to that, I will avoid the use of a splash screen

(when

Signed-off-by: ctigaret <cezar.tigaret@gmail.com>